pneumonia_detection/requirements.txt
streamlit==1.43.1
tensorflow==2.18.0
pillow==11.1.0
numpy==2.0.2
scikit-learn==1.6.1
scikeras==0.13.0
opencv-python==4.11.0.86
matplotlib==3.10.1
pandas==2.2.3
requests==2.32.3
